Added an executorid plugin
Currently marathon generates a unique ExecutorID for all tasks. This plugin allows to override this behavior. Given an application with a custom executor and a label with a key `MARATHON_EXECUTOR_ID` and a value of the executor Id it will use it to override the `TaskInfo.ExecutorInfo.ExecutorID`. See `executorid/README.md` for more details.